--- frontend/techpick-extension/src/pages/index.ts
@@ -1,2 +1,0 @@
-export { BookmarkPage } from './BookmarkPage';
-export { ErrorPage } from './ErrorPage';
